+++ src/Telegram.Bot/TelegramBotClient.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.IO;
 using System.Net;
 using System.Net.Http;
-using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
@@ -26,6 +25,8 @@ namespace Telegram.Bot
     /// </summary>
     public class TelegramBotClient : ITelegramBotClient
     {
+        private static readonly Update[] EmptyUpdates = { };
+
         private const string BaseUrl = "https://api.telegram.org/bot";
 
         private const string BaseFileUrl = "https://api.telegram.org/file/bot";
@@ -33,8 +34,6 @@ namespace Telegram.Bot
         private readonly string _baseRequestUrl;
 
         private readonly string _token;
-
-        private bool _invalidToken;
 
         private readonly HttpClient _httpClient;
 
@@ -157,9 +156,6 @@ namespace Telegram.Bot
         /// <exception cref="ArgumentException">Thrown if <paramref name="token"/> format is invalid</exception>
         public TelegramBotClient(string token, HttpClient httpClient = null)
         {
-            if (!Regex.IsMatch(token, @"^\d*:[\w\d-_]{35}$"))
-                throw new ArgumentException("Invalid token format", nameof(token));
-
             _token = token;
             _baseRequestUrl = $"{BaseUrl}{_token}/";
             _httpClient = httpClient ?? new HttpClient();
@@ -173,9 +169,6 @@ namespace Telegram.Bot
         /// <exception cref="ArgumentException">Thrown if <paramref name="token"/> format is invalid</exception>
         public TelegramBotClient(string token, IWebProxy webProxy)
         {
-            if (!Regex.IsMatch(token, @"^\d*:[\w\d-_]{35}$"))
-                throw new ArgumentException("Invalid token format", nameof(token));
-
             var httpClientHander = new HttpClientHandler
             {
                 Proxy = webProxy,
@@ -194,9 +187,6 @@ namespace Telegram.Bot
             IRequest<TResponse> request,
             CancellationToken cancellationToken = default)
         {
-            if (_invalidToken)
-                throw new ApiRequestException("Invalid token", 401);
-
             string url = _baseRequestUrl + request.MethodName;
 
             var httpRequest = new HttpRequestMessage(request.Method, url)
@@ -241,8 +231,6 @@ namespace Telegram.Bot
                 case HttpStatusCode.OK:
                     break;
                 case HttpStatusCode.Unauthorized:
-                    _invalidToken = true;
-                    throw new ApiRequestException("Invalid token", 401);
                 case HttpStatusCode.BadRequest when !string.IsNullOrWhiteSpace(responseJson):
                 case HttpStatusCode.Forbidden when !string.IsNullOrWhiteSpace(responseJson):
                 case HttpStatusCode.Conflict when !string.IsNullOrWhiteSpace(responseJson):
@@ -278,7 +266,8 @@ namespace Telegram.Bot
                 await GetMeAsync(cancellationToken).ConfigureAwait(false);
                 return true;
             }
-            catch (HttpRequestException e) when (e.Message.EndsWith("404 (Not Found)."))
+            catch (ApiRequestException e)
+                when (e.ErrorCode == 401)
             {
                 return false;
             }
@@ -293,9 +282,6 @@ namespace Telegram.Bot
         public void StartReceiving(UpdateType[] allowedUpdates = null,
             CancellationToken cancellationToken = default)
         {
-            if (_invalidToken)
-                throw new ApiRequestException("Invalid token", 401);
-
             _receivingCancellationTokenSource = new CancellationTokenSource();
 
             cancellationToken.Register(() => _receivingCancellationTokenSource.Cancel());
@@ -304,28 +290,24 @@ namespace Telegram.Bot
         }
 
 #pragma warning disable AsyncFixer03 // Avoid fire & forget async void methods
-        private async void ReceiveAsync(UpdateType[] allowedUpdates,
+        private async void ReceiveAsync(
+            UpdateType[] allowedUpdates,
             CancellationToken cancellationToken = default)
         {
             IsReceiving = true;
-
             while (!cancellationToken.IsCancellationRequested)
             {
                 var timeout = Convert.ToInt32(Timeout.TotalSeconds);
+                var updates = EmptyUpdates;
 
                 try
                 {
-                    var updates =
-                        await
-                            GetUpdatesAsync(MessageOffset, timeout: timeout, allowedUpdates: allowedUpdates,
-                                    cancellationToken: cancellationToken)
-                                .ConfigureAwait(false);
-
-                    foreach (var update in updates)
-                    {
-                        MessageOffset = update.Id + 1;
-                        OnUpdateReceived(new UpdateEventArgs(update));
-                    }
+                    updates = await GetUpdatesAsync(
+                       MessageOffset,
+                       timeout: timeout,
+                       allowedUpdates: allowedUpdates,
+                       cancellationToken: cancellationToken
+                    ).ConfigureAwait(false);
                 }
                 catch (OperationCanceledException)
                 {
@@ -337,6 +319,20 @@ namespace Telegram.Bot
                 catch (Exception generalException)
                 {
                     OnReceiveGeneralError?.Invoke(this, generalException);
+                }
+
+                try
+                {
+                    foreach (var update in updates)
+                    {
+                        OnUpdateReceived(new UpdateEventArgs(update));
+                        MessageOffset = update.Id + 1;
+                    }
+                }
+                catch
+                {
+                    IsReceiving = false;
+                    throw;
                 }
             }
 
@@ -928,7 +924,8 @@ namespace Telegram.Bot
             MakeRequestAsync(new EditInlineMessageTextRequest(inlineMessageId, text)
             {
                 DisableWebPagePreview = disableWebPagePreview,
-                ReplyMarkup = replyMarkup
+                ReplyMarkup = replyMarkup,
+                ParseMode = parseMode
             }, cancellationToken);
 
         /// <inheritdoc />