+++ test/Telegram.Bot.Tests.Integ/Inline Keyboard/CallbackQueryTests.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Threading.Tasks;
 using Newtonsoft.Json.Linq;
 using Telegram.Bot.Tests.Integ.Framework;
@@ -22,10 +22,9 @@ namespace Telegram.Bot.Tests.Integ.Inline_Keyboard
             _fixture = fixture;
         }
 
-        [Fact(DisplayName = FactTitles.ShouldReceiveAnswerCallbackQuery)]
+        [OrderedFact(DisplayName = FactTitles.ShouldReceiveAnswerCallbackQuery)]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendMessage)]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.AnswerCallbackQuery)]
-        [ExecutionOrder(1)]
         public async Task Should_Answer_With_Notification()
         {
             await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldReceiveAnswerCallbackQuery);
@@ -63,10 +62,9 @@ namespace Telegram.Bot.Tests.Integ.Inline_Keyboard
             ));
         }
 
-        [Fact(DisplayName = FactTitles.ShouldAnswerCallbackQueryWithAlert)]
+        [OrderedFact(DisplayName = FactTitles.ShouldAnswerCallbackQueryWithAlert)]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendMessage)]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.AnswerCallbackQuery)]
-        [ExecutionOrder(2)]
         public async Task Should_Answer_With_Alert()
         {
             await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldAnswerCallbackQueryWithAlert);